chore: Move language switcher types due to dts-generator
