+++ src/interfaces/languageSwitcherInterfaces.ts
